Comanager Bug fixed

+++ src/Presentation/Virgol.School/Controllers/Administrator/AdministratorController.cs
@@ -226,7 +226,7 @@ namespace Virgol.Controllers
                     password = model.MelliCode;
                 }
 
-                AdminDetail lastSchoolType = appDbContext.AdminDetails.OrderBy(x => x.SchoolsType).FirstOrDefault();
+                AdminDetail lastSchoolType = appDbContext.AdminDetails.OrderByDescending(x => x.SchoolsType).FirstOrDefault();
                 //newAdmin.UserType = Roles.Admin;
                 newAdmin.ConfirmedAcc = true;
                 IdentityResult result = await userManager.CreateAsync(newAdmin , password);
@@ -241,6 +241,10 @@ namespace Virgol.Controllers
                     AdminDetail adminDetail = new AdminDetail();
                     adminDetail.UserId = newAdmin.Id;
                     adminDetail.SchoolsType = model.schoolType;
+                    if(model.schoolType == 0)
+                        adminDetail.SchoolsType = lastSchoolType.SchoolsType + 1;
+
+                    adminDetail.TypeName = model.SchoolTypeName;
                     adminDetail.SchoolLimit = model.schoolLimit;
                     adminDetail.orgMoodleId = adminCatId;
 
@@ -310,14 +314,18 @@ namespace Virgol.Controllers
         {
             try
             {
+                UserModel userModel = appDbContext.Users.Where(x => x.Id == adminId).FirstOrDefault();
+
                 AdminDetail adminDetail = appDbContext.AdminDetails.Where(x => x.UserId == adminId).FirstOrDefault();
+                if(adminDetail != null)
+                {
+                    await moodleApi.DeleteCategory(adminDetail.orgMoodleId);
 
-                await moodleApi.DeleteCategory(adminDetail.orgMoodleId);
+                    appDbContext.Schools.RemoveRange(appDbContext.Schools.Where(x => x.SchoolType == adminDetail.SchoolsType).ToList());
+                    await appDbContext.SaveChangesAsync();
+                }
 
-                appDbContext.Schools.RemoveRange(appDbContext.Schools.Where(x => x.SchoolType == adminDetail.SchoolsType).ToList());
-                await appDbContext.SaveChangesAsync();
-
-                await userManager.DeleteAsync(appDbContext.Users.Where(x => x.Id == adminId).FirstOrDefault());
+                await UserService.DeleteUser(userModel);
 
 
                 return Ok(true);
